# flutter/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Travel Memory Vault API",
    "description": "Travel Memory Vault - Personal travel journal with AI-powered search",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Current User Profile",
        "description": "Get current user profile.\n    \n    **Authentication:** Required\n    \n    **Permissions:** Any authenticated user\n    \n    **Returns:**\n    User profile data\n    \n    **Response Example:**\n```json\n    {\n        \"id\": \"123e4567-e89b-12d3-a456-426614174000\",\n        \"email\": \"user@example.com\",\n        \"username\": \"traveler123\",\n        \"full_name\": \"John Doe\",\n        \"avatar_url\": \"https://example.com/avatar.jpg\",\n        \"bio\": \"Love to travel!\",\n        \"is_premium\": false,\n        \"is_verified\": true,\n        \"created_at\": \"2024-01-15T10:30:00Z\"\n    }\n```\n    \n    **Errors:**\n    - 401: Not authenticated\n    \n    **Business Logic:**\n    - Returns current user's profile data\n    - Used by frontend profile page",
        "operationId": "get_current_user_profile_api_v1_users_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__user__UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update Current User Profile",
        "description": "Update current user profile.\n    \n    **Authentication:** Required\n    \n    **Permissions:** Any authenticated user (can only update own profile)\n    \n    **Request Body:**\n    All fields are optional (partial update):\n    - username: New username (3-50 chars, alphanumeric + underscore)\n    - full_name: New full name\n    - bio: New bio (max 500 chars)\n    - avatar_url: New avatar URL\n    \n    **Returns:**\n    Updated user profile\n    \n    **Response Example:**\n```json\n    {\n        \"id\": \"123e4567-e89b-12d3-a456-426614174000\",\n        \"username\": \"new_username\",\n        \"full_name\": \"Updated Name\",\n        \"bio\": \"Updated bio\",\n        ...\n    }\n```\n    \n    **Errors:**\n    - 400: Username already taken or invalid format\n    - 401: Not authenticated\n    \n    **Business Logic:**\n    - Only updates provided fields (partial update)\n    - Username must be unique across all users\n    - Username can only contain letters, numbers, underscore\n    - Email cannot be changed (use Supabase Auth)",
        "operationId": "update_current_user_profile_api_v1_users_me_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__user__UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete Current User Account",
        "description": "Permanently delete current user account and owned data.\n\nDeletion sequence:\n1. Remove auth identity from Supabase Auth.\n2. Remove backend user row (DB cascades remove related rows).",
        "operationId": "delete_current_user_account_api_v1_users_me_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/stats": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Current User Stats",
        "description": "Get detailed user statistics.\n    \n    **Authentication:** Required\n    \n    **Permissions:** Any authenticated user\n    \n    **Returns:**\n    Detailed statistics about user's content and engagement\n    \n    **Response Example:**\n```json\n    {\n        \"trip_count\": 5,\n        \"place_count\": 47,\n        \"public_trip_count\": 2,\n        \"total_views\": 1234,\n        \"total_saves\": 56,\n        \"photos_uploaded\": 89\n    }\n```\n    \n    **Errors:**\n    - 401: Not authenticated\n    \n    **Business Logic:**\n    - Calculates statistics in real-time from database\n    - Used by dashboard to display user activity\n    - Premium users get additional stats (future)\n    \n    **Performance:**\n    - Uses SQLAlchemy aggregation functions\n    - Queries optimized with indexes\n    - Consider caching for high-traffic users",
        "operationId": "get_current_user_stats_api_v1_users_me_stats_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/profile": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Current User Complete Profile",
        "description": "Get complete user profile with statistics.\n    \n    **Authentication:** Required\n    \n    **Permissions:** Any authenticated user\n    \n    **Returns:**\n    User profile + statistics in single response\n    \n    **Response Example:**\n```json\n    {\n        \"user\": {\n            \"id\": \"123e4567-e89b-12d3-a456-426614174000\",\n            \"email\": \"user@example.com\",\n            \"username\": \"traveler123\",\n            ...\n        },\n        \"stats\": {\n            \"trip_count\": 5,\n            \"place_count\": 47,\n            ...\n        }\n    }\n```\n    \n    **Errors:**\n    - 401: Not authenticated\n    \n    **Business Logic:**\n    - Combines /me and /me/stats into single response\n    - Reduces frontend API calls for profile page\n    - More efficient than two separate requests",
        "operationId": "get_current_user_complete_profile_api_v1_users_me_profile_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/metadata": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Metadata",
        "operationId": "get_user_metadata_api_v1_users_me_metadata_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Upsert User Metadata",
        "operationId": "upsert_user_metadata_api_v1_users_me_metadata_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserMetadataUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserMetadataResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "dietary_restrictions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dietary Restrictions"
          },
          "budget_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Range"
          },
          "preferred_travel_style": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Travel Style"
          },
          "dislikes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dislikes"
          },
          "notification_enabled": {
            "type": "boolean",
            "title": "Notification Enabled"
          },
          "advisory_quiet_hours": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Quiet Hours"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "dietary_restrictions",
          "preferred_travel_style",
          "dislikes",
          "notification_enabled",
          "created_at",
          "updated_at"
        ],
        "title": "UserMetadataResponse"
      },
      "UserMetadataUpdate": {
        "properties": {
          "dietary_restrictions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dietary Restrictions"
          },
          "dislikes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dislikes"
          },
          "preferred_travel_style": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Travel Style"
          },
          "budget_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Range"
          },
          "notification_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Enabled"
          }
        },
        "type": "object",
        "title": "UserMetadataUpdate"
      },
      "UserProfileResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/app__schemas__user__UserResponse"
          },
          "stats": {
            "$ref": "#/components/schemas/UserStats"
          }
        },
        "type": "object",
        "required": [
          "user",
          "stats"
        ],
        "title": "UserProfileResponse",
        "description": "Complete user profile with stats.\n\nAttributes:\n    user: User profile data\n    stats: User statistics\n    \nUsed for dashboard/profile pages."
      },
      "UserStats": {
        "properties": {
          "trip_count": {
            "type": "integer",
            "title": "Trip Count",
            "default": 0
          },
          "place_count": {
            "type": "integer",
            "title": "Place Count",
            "default": 0
          },
          "public_trip_count": {
            "type": "integer",
            "title": "Public Trip Count",
            "default": 0
          },
          "total_views": {
            "type": "integer",
            "title": "Total Views",
            "default": 0
          },
          "total_saves": {
            "type": "integer",
            "title": "Total Saves",
            "default": 0
          },
          "photos_uploaded": {
            "type": "integer",
            "title": "Photos Uploaded",
            "default": 0
          }
        },
        "type": "object",
        "title": "UserStats",
        "description": "Detailed user statistics.\n\nAttributes:\n    trip_count: Total trips created\n    place_count: Total places across all trips\n    public_trip_count: Number of public trips\n    total_views: Total views across all trips\n    total_saves: Total saves across all trips\n    photos_uploaded: Number of photos uploaded\n    \nUsed by /users/me/stats endpoint for dashboard display."
      },
      "UserUpdate": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "User profile update schema.\n\nAll fields are optional for partial updates.\n\nAttributes:\n    username: New username (must be unique)\n    full_name: New full name\n    bio: New bio\n    avatar_url: New avatar URL\n    \nBusiness Rules:\n    - Username must be unique across all users\n    - Username can only contain alphanumeric and underscore\n    - Email cannot be changed (must use Supabase Auth)"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__user__UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "is_premium": {
            "type": "boolean",
            "title": "Is Premium"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified"
          },
          "subscription_ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Ends At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "username",
          "is_premium",
          "is_verified",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "User profile response schema.\n\nAttributes:\n    id: User UUID\n    email: User email\n    username: Unique username\n    full_name: Full name\n    avatar_url: Profile picture URL\n    bio: User bio\n    is_premium: Premium subscription status\n    is_verified: Email verification status\n    subscription_ends_at: Premium subscription end date\n    created_at: Account creation timestamp\n    last_login: Last login timestamp\n    \nConfig:\n    from_attributes: Enable ORM mode for SQLAlchemy models"
      }
    }
  }
}